+++ lib/src/header_parser/type_extractor/extractor.dart
@@ -73,6 +73,8 @@ Type getCodeGenType(Pointer<clang_types.CXType> cxtype, {String parentName}) {
       return Type.incompleteArray(
         clang.clang_getArrayElementType_wrap(cxtype).toCodeGenTypeAndDispose(),
       );
+    case clang_types.CXTypeKind.CXType_Bool:
+      return Type.boolean();
     default:
       if (cxTypeKindToSupportedNativeTypes.containsKey(kind)) {
         return Type.nativeType(